More feature flags fixes
